Moved StringLiteral fmt formatting to separate header
