+++ CMakeLists.txt
@@ -345,6 +345,7 @@ set(SIMPLNX_HDRS
   ${SIMPLNX_SOURCE_DIR}/Common/RgbColor.hpp
   ${SIMPLNX_SOURCE_DIR}/Common/ScopeGuard.hpp
   ${SIMPLNX_SOURCE_DIR}/Common/StringLiteral.hpp
+  ${SIMPLNX_SOURCE_DIR}/Common/StringLiteralFormatting.hpp
   ${SIMPLNX_SOURCE_DIR}/Common/TypeTraits.hpp
   ${SIMPLNX_SOURCE_DIR}/Common/Types.hpp
   ${SIMPLNX_SOURCE_DIR}/Common/TypesUtility.hpp
