+++ src/simplnx/Parameters/BoolParameter.cpp
@@ -1,5 +1,7 @@
 #include "BoolParameter.hpp"
 
+#include "simplnx/Common/StringLiteralFormatting.hpp"
+
 #include <fmt/core.h>
 #include <nlohmann/json.hpp>
