+++ src/simplnx/Parameters/CalculatorParameter.cpp
@@ -1,6 +1,7 @@
 #include "CalculatorParameter.hpp"
 
 #include "simplnx/Common/Any.hpp"
+#include "simplnx/Common/StringLiteralFormatting.hpp"
 #include "simplnx/Common/TypeTraits.hpp"
 #include "simplnx/DataStructure/BaseGroup.hpp"
 #include "simplnx/Utilities/SIMPLConversion.hpp"
