+++ src/simplnx/Parameters/DataPathSelectionParameter.cpp
@@ -1,6 +1,7 @@
 #include "DataPathSelectionParameter.hpp"
 
 #include "simplnx/Common/Any.hpp"
+#include "simplnx/Common/StringLiteralFormatting.hpp"
 
 #include <fmt/core.h>
 #include <nlohmann/json.hpp>
