+++ src/simplnx/Parameters/MultiArraySelectionParameter.cpp
@@ -1,6 +1,7 @@
 #include "MultiArraySelectionParameter.hpp"
 
 #include "simplnx/Common/Any.hpp"
+#include "simplnx/Common/StringLiteralFormatting.hpp"
 #include "simplnx/DataStructure/IArray.hpp"
 #include "simplnx/Utilities/SIMPLConversion.hpp"
 #include "simplnx/Utilities/StringUtilities.hpp"
