+++ src/simplnx/Parameters/NumberParameter.cpp
@@ -1,5 +1,7 @@
 #include "NumberParameter.hpp"
 
+#include "simplnx/Common/StringLiteralFormatting.hpp"
+
 #include <fmt/core.h>
 #include <nlohmann/json.hpp>
